split private header into two parts
It's better to split private header per each.
