+++ src/fw_iso_resource.c
@@ -1,11 +1,9 @@
 // SPDX-License-Identifier: LGPL-2.1-or-later
-#include "internal.h"
+#include "fw_iso_resource_private.h"
 
-#include <unistd.h>
 #include <sys/types.h>
 #include <sys/stat.h>
 #include <fcntl.h>
-#include <errno.h>
 #include <sys/ioctl.h>
 
 /**
